+++ spec/integration/requests/default_bot_spec.rb
@@ -1,0 +1,8 @@
+require 'integration_helper'
+
+RSpec.describe DefaultBotController, :telegram_bot, type: :request do
+  describe '#start' do
+    subject { -> { dispatch_command :start } }
+    it { should respond_with_message 'from default' }
+  end
+end